Consolidate cmdrunner fail-fast tests and expand coverage to 100%
